refactor(mode): move AppMode and TmuxActiveSet to src/app_mode.rs
