+++ src/input/default_mode.rs
@@ -14,8 +14,8 @@ use crate::input::ime::{ImeCommit, ImeState};
 use crate::input::shortcuts::{
     LeaderGate, LeaderPhase, LeaderStep, Shortcuts, clear_leader_phase, step_leader,
 };
+use crate::app_mode::AppMode;
 use crate::input::{InputPhase, current_modifiers};
-use crate::mode::AppMode;
 use crate::surface::OzmaTerminal;
 use crate::surface_geom::phys_to_pane_local;
 use crate::ui::copy_mode::{CopyModeState, EnterCopyModeActionEvent};
