+++ src/input/tmux/mouse.rs
@@ -13,10 +13,10 @@ mod effect;
 mod webview;
 
 use super::pane_hit::tmux_pane_at_phys;
+use crate::app_mode::TmuxActiveSet;
 use crate::configs::OzmuxConfigsResource;
 use crate::input::InputPhase;
 use crate::input::gesture::ClickTracker;
-use crate::mode::tmux::TmuxActiveSet;
 use crate::mode::tmux::copy_mode::{CopyModeSnapshot, cell_at_pane};
 use crate::mode::tmux::render::{DividerPixelRect, PackedTmuxLayout};
 use crate::ui::copy_mode::CopyModeState;
