+++ src/main.rs
@@ -1,6 +1,7 @@
 //! ozmux Bevy GUI entry point.
 
 mod action;
+mod app_mode;
 mod bootstrap;
 mod cef_profile;
 mod clipboard;
@@ -17,11 +18,11 @@ mod webview_pointer;
 mod window_title;
 
 use crate::action::ActionPlugin;
+use crate::app_mode::AppModePlugin;
 use crate::cef_profile::CefProfileDir;
 use crate::clipboard::ClipboardPlugin;
 use crate::input::focus::FocusSyncPlugin;
 use crate::input::hyperlink::HyperlinkInputPlugin;
-use crate::mode::AppMode;
 use crate::surface::SurfacePlugin;
 use crate::window_title::WindowTitlePlugin;
 use bevy::prelude::*;
@@ -51,9 +52,6 @@ fn main() {
     ensure_utf8_locale_env();
 
     let pre_configs = ozmux_configs::OzmuxConfigs::load().unwrap_or_default();
-    // The app boots into a single PTY shell; tmux is entered only by adopting the
-    // user's own `tmux -CC` (ControlModeDetected -> NextState(Tmux)), never at boot.
-    let initial_mode = AppMode::Default;
     let ozma_registry = WebviewAssetRegistry::default();
     let cef_profile = CefProfileDir::acquire().expect("create per-process CEF profile directory");
     App::new()
@@ -64,7 +62,7 @@ fn main() {
             }),
             cef_plugin(ozma_registry.clone(), cef_profile.path()),
         ))
-        .insert_state(initial_mode)
+        .add_plugins(AppModePlugin)
         .add_plugins((
             SurfacePlugin,
             DefaultModePlugin {
